Update CSS and HTML for functioning overlay when it is that players turn
The functioning overlay prevents the player from clicking on their own squares when it is their own turn. Implemented for the computers turn for consistent UX. Bug where the overlay includes the title of the player, when I change the CSS to 'player-grid game-board' the overlay applies to the body tag.