Updated class selector for ships and made the attempt X's red
New class selector allowed for the border of a computer ship to turn blue once it was hit. Addtionally I centered the X's from guesses. Need to find a way to increase the size of the X to the full box. Also need to include CSS and JS for when the guess is a miss.