Refactored various functions, added various functionality
Opponent and name attributes were useful for turn switching functionality. Moved iteration into registerGridDivEventListener function to prevent interaction with boards before the game had started. Added classToInteger function to assist with refactoring of HTMLtoboard function. Added functionality in HTMLtoboard function to check if a ship is sunk and to add a class for CSS, as well as to check if the game is over. Moved toggleOverlay into the Player class. Created triggerOverallOverlay function for gameover. Need to add functionality to overallOverlay to allow for the game to start again. BUG - Currently, only the final square of a sunken ship gets the appropriate class added to it. Want to implement a backtrack function to search for all nearby squares to fix the issue.