Added gameEndingOverlay div and first-player h3 to index file
Having gameEndingOverlay in the vanilla HTML allowed initialiseGame to be re-run once the game ended as the reset-btn can be identified when initialiseGame is run. Moving first-player into the vanilla HTML file prevented a bug which kept adding more first-player h3 components to the DOM on each reset of initialiseGame. With first-player in the HTML, the text content can be reset rather than writing separate code to add and remove it each time