Added functions for ship position stage
Changed draggable-ships to vertical to rearrange placement of ship container to LHS of player grid. Refactored button creation to seperate function for readability. Created removeShipsAndButtons and hideShips for logic for once the ship position stage ends and the gameplay stage commences. Created resetGrid and resetHTMLGrid to allow the logic for randomise to function correctly