Added resetShipsArray to resetGrid to prevent bugs where game end wouldn't trigger
As the ship array was not being reset after either the randomise or reset button was clicked, this resulted in more ships being in the array than were placed on the board. Accordingly the game could not end with the logic in place. As resetGrid is called whenever the randomise or reset button is clicked, the addition of resetShipsArray fixes this bug